Better compartmentalization and deletion handling
